requirement.txt
streamlit 
python-dotenv
pillow
google_generativeai

